calendar_backend/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1


djangorestframework==3.12.4

djongo==1.3.6
gunicorn==20.1.0
install==1.3.4
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
whitenoise==5.2.0
